requirements.txt
Flask
click
Jinja2
psutil
PyJWT
bcrypt
blinker
msgspec
requests
waitress
cachelib
Werkzeug
MarkupSafe
Flask-Cors
itsdangerous
Flask-Session
Flask-JWT-Extended
mysql-connector-python
